modified parallelization and test correction
Switched from using the threading package to the multiprocessing package for parallelization. Also converted the prep of the foundin data from serial parallel.

For multiple-test correction in addition to the B&H FDR now also compute a pseudo-Bonferroni where the nominal p-value is corrected but the number of traits tested for a QTL.