limited QTL checked and changed parallelization
Changed the parallelization from the threading to the multiprocessing package

Limited the QTL check for colocalization to features with a QTL detected by B&H FDR and the risk index variant has a QTL nominal p-value < 0.01. Before only limited on nominal pvalue but wanted to increase likelihood that actual QTL is present.