renamed tss to start distance
Newer version of tensorQTL outputs start_distance instead of tss_distance. Updated plot code that mad use of it.

For the day count compute added fillna after this on that column to deal with missing versus trying to apply to full dataframe before had which fails because of categorical columns.