reran based on DAn coloc
Reran the analysis with the flag turned on to limit the variants to the index variants for the loci that colocalize between PD risk and DAn meta-eQTL.

Increased the DPI values for the figures generated.

Updated the seaborn style used. Also replace the scale argument with the width_method argument for the seaborn boxenplot to correct the deprecations warning.